feat: gate refactored to reuse components into loyalty program feature
